+++ src/components/Header/Header.tsx
@@ -1,12 +1,8 @@
 import { LOGO } from '../../assets';
-import { useRecoilValue } from 'recoil';
-
-import { cartLengthSelector } from '../../recoil/myCartState';
+import CartButton from '../CartButton/CartButton';
 import * as Styled from './Header.styled';
 
 const Header = () => {
-  const cartLength = useRecoilValue(cartLengthSelector);
-
   return (
     <Styled.Header>
       <Styled.TitleDiv>
@@ -14,10 +10,7 @@ const Header = () => {
         <Styled.TitleHeading>SHOP</Styled.TitleHeading>
       </Styled.TitleDiv>
 
-      <Styled.CartButton>
-        <p>장바구니</p>
-        <Styled.CartLengthDiv>{cartLength}</Styled.CartLengthDiv>
-      </Styled.CartButton>
+      <CartButton />
     </Styled.Header>
   );
 };
